finished category routes

+++ models/Product.js
@@ -27,7 +27,7 @@ Product.init(
       },
     },
     stock: {
-      type: DataTypes.DECIMAL,
+      type: DataTypes.INTEGER,
       allowNull: false,
       defaultValue: 10,
       validate: {
@@ -35,7 +35,7 @@ Product.init(
       },
     },
     category_id: {
-      type: DataTypes.DECIMAL,
+      type: DataTypes.INTEGER,
       references: {
         model: 'category',
         key: 'id',
